Scene Transitions and Main Menu Created
Scene Transitions are made so scenes flow between each other a lot more smoothly now. Ain't that a sight to see, or I guess not see right? They're set up with the scene change button and there's a prefab for it now hooray!

Main Menu is set up and the points manager is also set up there too. You can click to go to the minigame but not back so that needs to be set up. You can also go to a rolling page but it's just the original back and forth to test.