Background Shader and Reveal Animation Finished
The background shader is created with each color and a lil shimmer to show what do what do and what rarity a character is. It looks sick but also we need to change the color per graph because the gradient thing is weird so I gotta do that I guess.

Reveal animation is done, I think i have to do it per character and I may soon crave death.