Characters can be rolled for now
Added a fancy new backdrop and implemented the prefabs for some characters. This is just so we can get a basic working thing to see the characters and stuff. It do work tho! The available character showing stuff is now funcitonal in the roll scene with a snazzy new background.